Extract NacDacPolicy unit tests to sibling file
Move the inline test module from nac_dac_policy.rs into
nac_dac_policy_tests.rs via #[path], preserving policy decisions
and fixture behavior.
